Authorization - user see only their boards - Add relation user-board to migations - Add relation user-board to model - Create new route - /user/boards, that display all boards belonging to the user - Move user login and registration to user-auth routes
