+++ hapi-mysql/lib/routes/users.js
@@ -1,77 +1,8 @@
 'use strict';
 
 const Joi = require('joi');
-const Helpers = require('./helpers');
-const User = require('../models/Users');
 
 module.exports = [
-    {
-        method: 'POST',
-        path: '/users/login',
-        options: {
-            description: 'Login',
-            notes: 'Login into user account',
-            tags: ['api'],
-            validate: {
-                payload: {
-                    name: Joi.string().required(),
-                    password: Joi.string().required()
-                }
-            },
-            handler: async (request, h) => {
-
-                const { name, password } = request.payload;
-                const { userService } = request.services();
-
-                const login = async (txn) => {
-
-                    return await userService.login({ name, password }, txn);
-                };
-
-                const user = await h.context.transaction(login);
-                const token = await userService.createToken(user.id);
-
-                return {
-                    user: { id: user.id, name: user.name },
-                    token
-                };
-            }
-        }
-    },
-    {
-        method: 'POST',
-        path: '/users/register',
-        options: {
-            description: 'Register',
-            notes: 'Register new user account',
-            tags: ['api'],
-            validate: {
-                payload: {
-                    name: Joi.string().required(),
-                    password: Joi.string().required()
-                }
-            },
-            handler: async (request, h) => {
-
-                const { userService } = request.services();
-
-                const signupAndFetchUser = async (txn) => {
-
-                    const id = await userService.signup(request.payload, txn);
-
-                    return await userService.findById(id, txn);
-                };
-
-                const user = await h.context.transaction(signupAndFetchUser);
-                const token = await userService.createToken(user.id);
-
-                return {
-                    user: { id: user.id, name: user.name },
-                    token
-                };
-            }
-        }
-    },
     {
         method: 'GET',
         path: '/user',
@@ -88,6 +19,23 @@ module.exports = [
                     user: { id: user.id, name: user.name },
                     token
                 };
+            }
+        }
+    },
+    {
+        method: 'GET',
+        path: '/user/boards',
+        options: {
+            description: 'Get current user boards',
+            notes: 'Returns current user boards, require token',
+            tags: ['api'],
+            auth: 'jwt',
+            handler: async (request) => {
+
+                const { credentials: user } = request.auth;
+                const { Boards } = request.models();
+
+                return await Boards.query().where('user_id', user.id);
             }
         }
     }
